+++ src/main/java/com/andrey/todolist/service/ToDoItemService.java
@@ -10,9 +10,9 @@ public interface ToDoItemService {
     public Iterable<ToDoItem> findAllByUser(Principal principal);
     public ToDoItem addToDoItem(ToDoItem toDoItem, Principal principal);
     public ToDoItem editToDo(ToDoItem newToDoItem, Principal principal);
-    public void deleteToDo(int id, Principal principal);
-    public void completeToDo(int id, Principal principal);
-    public boolean toDoExists(int id);
-    public ToDoItem findToDoItemById(int id);
+    public void deleteToDo(Long id, Principal principal);
+    public void completeToDo(Long id, Principal principal);
+    public boolean toDoExists(Long id);
+    public ToDoItem findToDoItemById(Long id);
     public boolean canUserAccessToDo(ToDoItem toDoItem, User user);
 }
